Refactor provider tests so we can export utilities
Exporting the utilities needed to run provider tests should allow
implementers to run their own tests using responses they can't share
upstream.
