+++ providertests/testdata/auth0_cert.pem
