+++ providertests/utils.go
@@ -1,4 +1,4 @@
-package saml2
+package providertests
 
 import (
 	"crypto/x509"
@@ -6,9 +6,8 @@ import (
 	"encoding/pem"
 	"io/ioutil"
 	"testing"
-	"time"
 
-	"github.com/jonboulle/clockwork"
+	"github.com/russellhaering/gosaml2"
 	"github.com/russellhaering/goxmldsig"
 	"github.com/stretchr/testify/require"
 )
@@ -45,23 +44,10 @@ func LoadCertificateStore(path string) dsig.X509CertificateStore {
 
 type ProviderTestScenario struct {
 	Response        string
-	ServiceProvider *SAMLServiceProvider
+	ServiceProvider *saml2.SAMLServiceProvider
 }
 
-func TestValidateResponses(t *testing.T) {
-	scenarios := []ProviderTestScenario{
-		{
-			Response: LoadXMLResponse("./testdata/auth0_response.xml"),
-			ServiceProvider: &SAMLServiceProvider{
-				IdentityProviderSSOURL:      "https://scaleft-test.auth0.com/samlp/rlXOZ4kOUTQaTV8icSXrfZUd1qtD1NhK",
-				IdentityProviderIssuer:      "urn:scaleft-test.auth0.com",
-				AssertionConsumerServiceURL: "http://localhost:8080/v1/_saml_callback",
-				IDPCertificateStore:         LoadCertificateStore("./testdata/auth0_cert.pem"),
-				Clock:                       dsig.NewFakeClock(clockwork.NewFakeClockAt(time.Date(2016, 7, 25, 17, 50, 0, 0, time.UTC))),
-			},
-		},
-	}
-
+func ExerciseProviderTestScenarios(t *testing.T, scenarios []ProviderTestScenario) {
 	for _, scenario := range scenarios {
 		_, err := scenario.ServiceProvider.RetrieveAssertionInfo(scenario.Response)
 		require.NoError(t, err)
